- update to libsbml 5.12

+++ src/libSBML/src/sbml/conversion/SBMLConverterRegistry.cpp
@@ -94,12 +94,11 @@ SBMLConverterRegistry::getConverterByIndex(int index) const
 {
   if (index < 0 || index >= getNumConverters())
     return NULL;
-  return mConverters.at(index)->clone();
+  return mConverters.at((size_t)index)->clone();
 }
 
 
 /** @cond doxygenLibsbmlInternal */
-
 SBMLConverterRegistry::SBMLConverterRegistry()
 {
 }
@@ -119,7 +118,6 @@ SBMLConverterRegistry::~SBMLConverterRegistry()
   }
   mConverters.clear();
 }
-
 /** @endcond */
