+++ src/libSBML/src/sbml/packages/layout/sbml/BoundingBox.cpp
@@ -668,20 +668,20 @@ void BoundingBox::readAttributes (const XMLAttributes& attributes,
   if (getErrorLog() != NULL)
   {
     numErrs = getErrorLog()->getNumErrors();
-    for (int n = numErrs-1; n >= 0; n--)
+    for (int n = (int)numErrs-1; n >= 0; n--)
     {
-      if (getErrorLog()->getError(n)->getErrorId() == UnknownPackageAttribute)
+      if (getErrorLog()->getError((unsigned int)n)->getErrorId() == UnknownPackageAttribute)
       {
         const std::string details =
-          getErrorLog()->getError(n)->getMessage();
+          getErrorLog()->getError((unsigned int)n)->getMessage();
         getErrorLog()->remove(UnknownPackageAttribute);
         getErrorLog()->logPackageError("layout", LayoutBBoxAllowedAttributes,
           getPackageVersion(), sbmlLevel, sbmlVersion, details, getLine(), getColumn());
       }
-      else if (getErrorLog()->getError(n)->getErrorId() == UnknownCoreAttribute)
+      else if (getErrorLog()->getError((unsigned int)n)->getErrorId() == UnknownCoreAttribute)
       {
         const std::string details =
-          getErrorLog()->getError(n)->getMessage();
+          getErrorLog()->getError((unsigned int)n)->getMessage();
         getErrorLog()->remove(UnknownCoreAttribute);
         getErrorLog()->logPackageError("layout", 
           LayoutBBoxAllowedCoreAttributes,
@@ -738,9 +738,7 @@ BoundingBox::getTypeCode () const
 }
 
 
-/*
- * Accepts the given SBMLVisitor.
- */
+/** @cond doxygenLibsbmlInternal */
 bool
 BoundingBox::accept (SBMLVisitor& v) const
 {
@@ -753,6 +751,7 @@ BoundingBox::accept (SBMLVisitor& v) const
   
   return true;
 }
+/** @endcond */
 
 /*
  * Creates an XMLNode object from this.
@@ -812,7 +811,6 @@ BoundingBox::enablePackageInternal(const std::string& pkgURI,
 
 #endif /* __cplusplus */
 /** @cond doxygenIgnored */
-
 LIBSBML_EXTERN
 BoundingBox_t *
 BoundingBox_create (void)
@@ -1032,7 +1030,6 @@ BoundingBox_unsetId (BoundingBox_t *bb)
   if (bb == NULL) return;
   bb->unsetId();
 }
-
 /** @endcond */
 LIBSBML_CPP_NAMESPACE_END
